SCENARIOS: Added state machine lib, trigger script
Added the MonsterLove state machine lib we'll be using
for our scenarios. I made some minor mods to it, namely
modded it to support the "OnTriggerEnter" callback as a
trigger for state transitions and a "ChangeState" hook
that takes in a trigger id as an arg

Added a template for how a scenario script will look - these
scripts will be attached to an empty game object in your scene

Added a generic "trigger" script that can be attached to any
gameobject with a triggerable collider, to be used to initiate
a state transition (e.g. it calls your scenarios `ChangeState(id)`
method)