HUD-DEVELOP: made public fields private
Also some thoughts on updating text...

I think the best way to do this is not to add methods to the controller itself; but rather, when needing to change/update text, to switch out the model used by the controller. For example, if an event is triggered which requires the HUD text to be changed, replace the existing HUDModel with a new one that has the desired text. This way, the controller itself does not maintain state, state can be encapsulated within HUDModels, and scenarios do not have to store text/state either other than to update HUDController with a new HUDModel.