SPAWNCONTROLLER-DEV: cleaned up SpawnController
cleaned up the code so that it is better.
now it is very good code.
also the code uses a SpawnModel which has inital spawn objects, and on demand spawn objects.
initial spawn objects are spawned onScenarioEnter(), and on demand spawn objects are spawned by index
you don't have to store your on demand spawns in a SpawnModel, but it is helpful if you want to keep your scenario class clean.